fix link filtering

+++ src/modules/linkFilter.js
@@ -1,6 +1,6 @@
 const { readdirSync, readdir } = require('fs')
 const ExtendedClient = require('../class/ExtendedClient')
-const { Message } = require('discord.js')
+const { Message, EmbedBuilder } = require('discord.js')
 
 /**
  * 
@@ -9,12 +9,24 @@ const { Message } = require('discord.js')
  */
 
 module.exports = (bot, message) => {
+
+    async function sendMessage(message){
+        const embed = new EmbedBuilder()
+            .setColor('red')
+            .setDescription(message)
+    }
     const inviteRegex = /(discord\.(gg|com|io|me|app\/invite)\/.+|discordapp\.com\/invite\/.+)/gi;
     const logChannel = bot.channels.cache.get(process.env.MESSAGE_CHANNEL)
+    const allowed = [
+        "473568312840814610",
+        "274021702411747328"
+    ]
+
+    if(allowed.includes(message.author.id)) return;
 
     if (inviteRegex.test(message.content)) {
         message.delete()
-            .then(msg => logChannel.send(`Deleted message from ${msg.author.username}: ${msg}`))
+            .then(msg => logChannel.send(sendMessage(`Deleted message from ${msg.author.username}: ${msg}`)))
             .catch(console.error);
         
         message.channel.send(`${message.author}, sending Discord invite links is not allowed here!`)
